Pizza shop files
This is a beginner-level project aimed at providing an introduction to SQL and TABLEAU. The dataset for this project is sourced from Kaggle. I hope this project proves beneficial! I am open to feedback or any assistance you may require. Please feel free to reach out to me, and I will do my best to help.